ZERO-161 [Rename]: UI 하위 폴더로 해당 파일 이동
