+++ src/components/tasks/UI/modal/AddTaskListModal.tsx
@@ -8,7 +8,7 @@ import { TASKLIST_REQUEST_INIT } from '@constants/initValues';
 import { useTaskListStore } from '@/src/stores/taskListStore';
 import { postTaskList, TaskListUrlParams } from '@/src/api/tasks/taskListAPI';
 import type { TaskListRequestBody } from '@/src/types/tasks/taskListDto';
-import NameInput from './input/NameInput';
+import NameInput from '../input/NameInput';
 
 interface AddTaskListModalProps {
   isOpen: boolean;
